Moved ceph cluster specific tests to new file cephcluster_test.go
Signed off by : Priyanka Jiandani <pjiandan@redhat.com>
